+++ app/sql_app/models.py
@@ -4,7 +4,6 @@ import datetime
 import os
 import sys
 sys.path.append(os.path.dirname(os.path.abspath(os.path.dirname(os.path.abspath(os.path.dirname(__file__))))))
-import database
 from database import Base
 from sqlalchemy.sql import func
 sys.path.append(os.path.dirname(os.path.abspath(os.path.dirname(__file__))))
